fix: Add a Makefile
I do not have time to be hunting build commands around, particularly
those buried deep down in GitHub Action workflow files.

This is very basic: build the thing. It tries to avoid rebuilding stuff
if not necessary, but it's not perfect. Run `make clean && make build`
if you switch versions.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>