feat: Normalize build system (#170)
The other day I was trying to validate something simple, and I was
surprised to find that it cannot be done because the only place in the
code where that particular process runs is inside a Github workflow.

The problem with that is that it makes local validation error-prone and
unnecessarely complex: you have to know that the process runs in a
workflow; you have to copy-paste the commands from the workflow into
something else; if you are reasonable, "something else" is a script,
which means that now you have a script that is not part of the
repository running a process that is part of the repository, which opens
the window for the local thing to divergue from the official thing; ...

All those problems do not exist if the workflow runs something that is
part of the repository *and* that can be run locally. Since I'm not here
to reinvent the wheel, drop a makefile in here that is already doing all
those things are more (e.g. drop a golangci-lint configuration file so
that the lint step actually runs instead of getting skipped).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>